Move the cudnn module.
